Adjusted zigzag
I tried to adjust zigzag so that it now correctly orders the components. The DWT_DCT is still performing with random guessing accuracy. I will now move on to the PCA_CNN mapping, and come back to DWT_DCT and perform thorough debugging to find the major mistakes (it may be that there are other problems that are destroying performance, early on).