fixed DWT_DCT_SVM, starting PCA
DWT_DCT_SVM wasn't working due to the fact the training data and its labels were not properly matched. I fixed this issue. Also, I have tried a new normalizing scheme from 0-1. The results of which can be seen in 'results save.txt'. The long and short is that DWT_DCT performance is unaffected by this. performance for DWT_DCT is anywhere from 88-92% depending on features. Interestingly, this is very much not coherent with the original paper. I will revisit DWT_DCT_SVM once I have PCA_CNN working and try to optimize. If optimizing is not possible then I will do statistical anlysis on the current data and compare the two. Provided that PCA_CNN and DWT_DCT_SVM are primarily of interest for efficiency, a lower than reported value will still hold value in this regard.