got DWT_DCT to 95% and started CNN
for DWT_DCT I changed gamma to 'scale' and got 95% fro 60k and 60 features. which is about equal to the paper. tunning gamma some more will likely result in ideal performance.

PCA_CNN:
I created a PCA document. I followed the paper exactly. I then tested it on a basic SVM to check fi the data is valid. Initially this did not work (10% classificaiton); however, as it would turn out this is because of the fact that you must perform PCA on ALL the data, not seperate.